Cleanup

+++ src/fin_vm.c
@@ -98,7 +98,7 @@ void fin_vm_interpret(fin_ctx* ctx, fin_mod_func* func, fin_val* stack) {
             FIN_VM_NEXT();
         }
         FIN_VM_OP(fin_op_branch) {
-            int16_t offset = *ip++;
+            int32_t offset = *ip++;
             offset |= *ip++ << 8;
             ip += offset;
             FIN_VM_NEXT();
@@ -107,7 +107,7 @@ void fin_vm_interpret(fin_ctx* ctx, fin_mod_func* func, fin_val* stack) {
             if ((--top)->b)
                 ip += 2;
             else {
-                int16_t offset = *ip++;
+                int32_t offset = *ip++;
                 offset |= *ip++ << 8;
                 ip += offset;
             }
